Refactor mock
