+++ mock/data/statuskort/content.ts
@@ -1,0 +1,13 @@
+import type { Locale } from "./utils";
+
+export const tittel: Record<Locale, string> = {
+  nb: "Arbeidsavklaringspenger",
+  nn: "Arbeidsavklaringspenger",
+  en: "Arbeidsavklaringspenger",
+};
+
+export const beskrivelse: Record<Locale, string> = {
+  nb: "Vi behandler søknaden din",
+  nn: "Vi behandler søknaden din",
+  en: "Vi behandler søknaden din",
+};
