+++ mock/data/statuskort/utils.ts
